Add 3rem margin-bttom to Alert box
The issue was there is no space between bottom of alert box and following content, looking crammed in rendering. The solution is to add 3 rem margin-bottom to class usa-alert. The change will impact all md documents that have Alert box; We need to watch for regression in existing document rendering.