removing duplication
I started edits to push through to the piv-guides repo
I'm going to let these edits go in via the pull requests; then start another round.
In general, it seemed that the "template.md" is trying to be the Contribute page?  With all the same info plus a lot more?  Is it necessary? 

I will update contribute, contribute_addpage and contribute_editpage with the info instead.